Changed transaction action settings default values.
Now that classes are not mapped on runtime, the only missing setting is
the number of attempts, so change the default value and the
documentation to reflect the change.

Not adding a new sql query to update the setting for exercise attempt
transactions since the values there will simply not be used.