New structure for the ckeditor dependencies see #5875
The new structure allows the use of other editors such as TinyMce, that
plays well with elfinder. The dependencies were moved inside
ChamiloLMS/Component/Editor classes. Now the editor is used as a
service. The problem with ckeditor is that, it does not works properly
in smartphones/tablets event with the latest version 4.3.