htaccess: Add install step for a proper .htaccess
Install a new writeHtaccess() function in the install process that takes
/htaccess-dist and replaces {BASE_URL} for the appropiate value used by
RewriteBase to find Chamilo files.

This solves the problem of Chamilo installations that are not hosted on
a web server root. Namely, all the http://domain/chamilo-lms installs.

The method follows similar practice by WordPress where the .htaccess
file is generated by WordPress on demand when the user activates pretty
permalinks.

A follow–up commit enhances the safety of this process by making sure
the site URL is clean enough for RewriteBase.