install: Clean trailing slashes in site URL
This ensures that accidental trailing slashes does not break the site
after installation.

This is important for the automatic generation of the .htaccess file,
since RewriteBase depends on a valid, clean, URL path to work.

When RewriteBase is wrong, many different and obscure errors can pop in
the site, so it's better to be very sure.

Arguably, this might not be the best place for input filtering during
the installation.